requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
httpx
openai
python-multipart
